Update README.md
Delete "support Ukraine"! We outside of politics and not support any wars. Because the discussion of politics and peace, as well as the beginning and end of wars, the collection of assistance to one or another side of the conflict, should take place on other specialized platforms. Embedding one point of view everywhere in the form of propaganda leads to an escalation of nationalism and support for the existing war, which will definitely not lead to the peaceful coexistence of the various peoples of our planet.